add PinLogin refactor into smaller components.
